feat(alphamap): extract the alpha hitmask baker into its own package
Each flatland-bake baker is meant to be a small grab-it-when-you-need-it
package, not a tenant of an unrelated one. The alpha baker had been
parked in @three-flatland/normals (which already owned PNG decode), but
the name was misleading — normals now shipped a non-normal baker.

Move the alpha baker into a new @three-flatland/alphamap package: the
`flatland-bake alpha` subcommand, bakeAlphaMapFile, and the
ALPHA_DESCRIPTOR. The runtime side (AlphaMap, resolveAlphaMap) stays in
three-flatland/events, decoupled by the re-declared descriptor literal —
no cross-package import either way. normals goes back to a single
`normal` baker.

Shared deps move to the workspace catalog (pngjs, @types/pngjs); both
baker packages reference catalog:. Discovery is unchanged — the bake CLI
finds whichever bakers are installed via their package.json flatland.bake
manifest, so consumers install only the baker they need.
